Refactor specs to use in-memory SQLite and shoulda-matchers (#9)
Release v0.5.0

Part of LG-4095, removing duplicate code and specs from the IdP and into
the gem. While doing so it seemed to make sense to clean things up a bit
so we added the in-memory SQLite trick from 18F/identity-idp-config to
use normal ActiveRecord and added Shoulda Matchers. The cert specs were
a little more involved so we left those alone for now.